refactor: 更新属性处理逻辑，优化能力值存储格式并重构工具函数
