+++ app/components/common/dicer/cmdType.d.ts
@@ -77,7 +77,9 @@ interface RoleAbility { // eslint-disable-line ts/consistent-type-definitions
   roleId?: number;
   ruleId?: number;
   act?: Record<string, string>;
-  ability?: Record<string, number>;
+  basic?: Record<string, string>;
+  ability?: Record<string, string>;
+  skill?: Record<string, string>;
 }
 
 interface UserRole { // eslint-disable-line ts/consistent-type-definitions
